backend/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-cors==4.0.0
flask-jwt-extended==4.5.2
flask-restx==1.1.0
Flask-Migrate==4.0.5
Flask-Bcrypt==1.0.1
psycopg2-binary==2.9.7
redis==4.6.0
python-dotenv==1.0.0
bcrypt==4.0.1
marshmallow==3.20.1
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
reportlab==4.0.4
fpdf2==2.7.4
openpyxl==3.1.2
gunicorn==21.2.0
requests==2.31.0
python-dateutil==2.8.2